refactor: OrderTable 바깥으로 이동
