+++ src/main/java/kitchenpos/order/order/domain/OrderRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.order.order.domain;
 
-import kitchenpos.order.eatin.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTable;
 
 import java.util.List;
 import java.util.Optional;
