+++ src/main/java/kitchenpos/ordertable/application/OrderTableService.java
@@ -1,9 +1,9 @@
-package kitchenpos.order.eatin.application;
+package kitchenpos.ordertable.application;
 
 import kitchenpos.order.order.domain.OrderRepository;
 import kitchenpos.order.order.domain.OrderStatus;
-import kitchenpos.order.eatin.domain.OrderTable;
-import kitchenpos.order.eatin.domain.OrderTableRepository;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
