+++ src/main/java/kitchenpos/ordertable/ui/OrderTableRestController.java
@@ -1,7 +1,7 @@
-package kitchenpos.order.eatin.ui;
+package kitchenpos.ordertable.ui;
 
-import kitchenpos.order.eatin.application.OrderTableService;
-import kitchenpos.order.eatin.domain.OrderTable;
+import kitchenpos.ordertable.application.OrderTableService;
+import kitchenpos.ordertable.domain.OrderTable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
