+++ src/test/java/kitchenpos/application/InMemoryOrderTableRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.application;
 
-import kitchenpos.order.eatin.domain.OrderTable;
-import kitchenpos.order.eatin.domain.OrderTableRepository;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTableRepository;
 
 import java.util.*;
